requirements.txt
# Core dependencies for the RAG application
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
starlette>=0.45.0
jinja2>=3.1.4
python-multipart>=0.0.9

# Vector database
weaviate-client>=4.6.0
chromadb>=1.0.0

# AI and ML
openai>=1.80.0
sentence-transformers>=4.0.0
transformers>=4.50.0
torch>=2.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Document processing
pdfplumber>=0.10.0
pypdf>=5.0.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0

# HTTP and networking
httpx>=0.27.0
requests>=2.32.0
cachetools>=5.3.0

# Data validation and serialization
pydantic>=2.11.0
python-dotenv>=1.1.0

# JWT Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# MongoDB
pymongo>=4.6.0

# File operations
aiofiles>=24.1.0

# Compatible protobuf version (resolves conflicts)
protobuf>=5.0.0,<6.0.0

# Development and utilities
python-dateutil>=2.9.0
PyYAML>=6.0.0
tqdm>=4.67.0
click>=8.2.0
tiktoken>=0.7.0

# Speech Integration Dependencies
openai-whisper==20250625
edge-tts==7.0.2
pygame==2.6.1
pyaudio==0.2.14

